Fix a portion of warnings, organize source code, remove unnecessary semicolons
